requirements.txt
requests~=2.31.0
Flask~=2.3.2
SQLAlchemy~=2.0.25
WTForms~=3.0.1
python-dotenv~=1.0.1